+++ src/Console/Commands/InstallCommand.php
@@ -34,5 +34,7 @@ class InstallCommand extends Command
         $this->call('cortex:publish:auth', ['--force' => $this->option('force'), '--resource' => $this->option('resource')]);
         $this->call('cortex:migrate:auth', ['--force' => $this->option('force')]);
         $this->call('cortex:seed:auth');
+
+        $this->call('cortex:activate', ['--module' => 'cortex/auth']);
     }
 }